Removing codegen for PMM
