+++ cs/src/core/FASTER.core.csproj
@@ -44,9 +44,9 @@
     <PackageReference Include="System.Runtime.Loader" Version="4.3.0" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Index/FASTER/AsyncIO.cs" />
     <EmbeddedResource Include="Index/FASTER/Checkpoint.cs" />
     <EmbeddedResource Include="Index/Common/Contexts.cs" />
+    <EmbeddedResource Include="Index/FASTER/AsyncIO.cs" />
     <EmbeddedResource Include="Index/FASTER/FASTER.cs" />
     <EmbeddedResource Include="Index/FASTER/FASTERBase.cs" />
     <EmbeddedResource Include="Index/FASTER/FASTERImpl.cs" />
@@ -65,7 +65,6 @@
     <EmbeddedResource Include="ManagedLayer/MixedUnwrappedTypes.cs" />
     <EmbeddedResource Include="ManagedLayer/MixedUserFunctions.cs" />
     <EmbeddedResource Include="ManagedLayer/MixedValueWrapper.cs" />
-    <EmbeddedResource Include="Allocator/PersistentMemoryMalloc.cs" />
     <EmbeddedResource Include="Utilities/AsyncResultTypes.cs" />
     <EmbeddedResource Include="Utilities/StateTransitions.cs" />
   </ItemGroup>
